extracted modules sources and RuleFromTslint
